Rename SpaceIndex to Space
