+++ src/views/Space.vue
